app/requirements.txt
discord.py
gTTS
pynacl